multiOFF/requirements.txt
torch==2.0.1
transformers==4.30.2
pillow==9.5.0
pandas==2.0.2
wandb==0.15.4
numpy==1.24.3
tqdm==4.65.0
scikit-learn==1.2.2
matplotlib==3.7.1 